--- .env
@@ -1,7 +1,0 @@
-VITE_LOCAL_ADDRESS=http://localhost:5173
-VITE_SERVER_ADDRESS=https://kkoalla.app:8443
-
-# VITE_REDIRECT_URI =
-#   process.env.NODE_ENV === "development"
-#     ? "http://localhost:3000/auth/callback"
-#     : "https://myapp.com/auth/callback";
